feat: full 1.19.3 support
ok this is SO weird. protocollib appears to have absolutly NO trouble modifing (via reflection) a IMMUTABLE list??
the code is so messy right now but i'll clean it up later, let's say it works for now.
this commit also breaks all the other versions due to the way i get the internals soooooo.....